# api-spec.yaml
openapi: 3.0.3
info:
  title: Next Loyalty Program API
  description: 'Unified Loyalty Program API for Next plc omnichannel retail.

    Supports customer registration, points earning/redemption, tier management,

    notifications, staff operations, and admin configuration.

    '
  version: 1.0.0
  contact:
    name: Next Loyalty Engineering
    email: loyalty-engineering@next.co.uk
servers:
- url: https://loyalty.next.co.uk/api/v1
  description: Production
- url: https://staging.loyalty.next.co.uk/api/v1
  description: Staging
- url: http://localhost:3000/api/v1
  description: Local Development
security:
- bearerAuth: []
...
paths:
...
  /customers/identify:
    get:
      operationId: identifyCustomer
      summary: Identify customer at POS
      description: Look up a customer by phone, loyalty ID, or QR code for in-store
        identification. Requires API key auth (POS terminal).
      tags:
      - Customers
      security:
      - apiKeyAuth: []
      parameters:
      - name: method
        in: query
        required: true
        schema:
          type: string
          enum:
          - phone
          - loyalty_id
          - qr
      - name: value
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Customer identified
          content:
            application/json:
              schema:
                type: object
                properties:
                  customerId:
                    type: string
                    format: uuid
                  name:
                    type: string
                  tier:
                    type: string
                    example: Gold
                  pointsBalance:
                    type: integer
                    example: 2500
                  memberSince:
                    type: string
                    format: date
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /points/earn:
    post:
      operationId: earnPoints
      summary: Earn points on purchase
      description: Credit points to a customer based on a purchase transaction. Idempotent
        by transaction reference. Requires API key (POS/e-commerce system).
      tags:
      - Points
      security:
      - apiKeyAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - customerId
              - transactionAmount
              - referenceId
              - channel
              properties:
                customerId:
                  type: string
                  format: uuid
                transactionAmount:
                  type: number
                  format: double
                  minimum: 0.01
                  example: 75.5
                referenceId:
                  type: string
                  description: Order or transaction ID from source system
                  example: ORD-2026-001234
                channel:
                  type: string
                  enum:
                  - online
                  - in_store
                  - app
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      sku:
                        type: string
                      category:
                        type: string
                      amount:
                        type: number
                        format: double
      responses:
        '201':
          description: Points earned
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactionId:
                    type: string
                    format: uuid
                  pointsEarned:
                    type: integer
                    example: 76
                  bonusPoints:
                    type: integer
                    example: 0
                  newBalance:
                    type: integer
                    example: 1576
                  campaignApplied:
                    type: string
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /points/redeem:
    post:
      operationId: redeemPoints
      summary: Redeem points at checkout
      description: Deduct points and apply discount to an order. Validates balance,
        minimum threshold, and max discount rules. Atomic operation with row-level
        locking.
      tags:
      - Points
      security:
      - bearerAuth: []
      - apiKeyAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - pointsToRedeem
              - orderId
              - orderTotal
              - channel
              properties:
                pointsToRedeem:
                  type: integer
                  minimum: 1
                  example: 500
                orderId:
                  type: string
                  example: ORD-2026-005678
                orderTotal:
                  type: number
                  format: double
                  example: 120.0
                channel:
                  type: string
                  enum:
                  - online
                  - in_store
                  - app
      responses:
        '200':
          description: Points redeemed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  transactionId:
                    type: string
                    format: uuid
                  pointsRedeemed:
                    type: integer
                    example: 500
                  discountApplied:
                    type: number
                    format: double
                    example: 25.0
                  remainingBalance:
                    type: integer
                    example: 1076
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /points/balance:
    get:
      operationId: getPointsBalance
      summary: Get points balance
      description: Retrieve the authenticated customer's current points balance with
        monetary equivalent and tier multiplier.
      tags:
      - Points
      responses:
        '200':
          description: Points balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PointsBalance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /points/transactions:
    get:
      operationId: getPointsTransactions
      summary: Get points transaction history
      description: Paginated list of all points transactions for the authenticated
        customer. Filterable by type and date range.
      tags:
      - Points
      parameters:
      - $ref: '#/components/parameters/PageNumber'
      - $ref: '#/components/parameters/PageSize'
      - name: type
        in: query
        schema:
          type: string
          enum:
          - earn
          - redeem
          - adjustment
          - bonus
          - expired
          - reversal
      - name: fromDate
        in: query
        schema:
          type: string
          format: date
      - name: toDate
        in: query
        schema:
          type: string
          format: date
      responses:
        '200':
          description: Transaction list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /customers/me/qr-code:
    get:
      operationId: getCustomerQrCode
      summary: Generate dynamic QR code for POS identification
      description: 'Returns a time-limited encrypted QR code payload containing the
        customer''s

        loyalty identifier. QR refreshes every 60 seconds. Used by store associates

        to scan and identify the customer at POS.

        '
      tags:
      - Customers
      security:
      - bearerAuth: []
      responses:
        '200':
          description: QR code payload generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      qrPayload:
                        type: string
                        description: Encrypted payload for QR code rendering
                      expiresAt:
                        type: string
                        format: date-time
                      refreshInSeconds:
                        type: integer
                        example: 60
        '401':
          $ref: '#/components/responses/Unauthorized'
  /customers/me/dashboard:
    get:
      operationId: getCustomerDashboard
      summary: Get aggregated loyalty dashboard data
      description: 'Returns all data needed to render the customer home screen in
        a single call:

        points balance, tier info, progress to next tier, recent transactions, and

        active promotions. BFF-optimised endpoint to reduce mobile round-trips.

        '
      tags:
      - Customers
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Dashboard data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      balance:
                        $ref: '#/components/schemas/PointsBalance'
                      tier:
                        $ref: '#/components/schemas/Tier'
                      nextTierProgress:
                        type: object
                        properties:
                          nextTierName:
                            type: string
                          pointsRequired:
                            type: integer
                          pointsEarned:
                            type: integer
                          progressPercent:
                            type: number
                      recentTransactions:
                        type: array
                        maxItems: 5
                        items:
                          $ref: '#/components/schemas/Transaction'
                      activePromotions:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                            name:
                              type: string
                            description:
                              type: string
                            endsAt:
                              type: string
                              format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (15-minute expiry)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for POS terminals and e-commerce integrations
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Unique key to ensure idempotent processing of write operations
    PageNumber:
      name: pageNumber
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
  schemas:
...
    PointsBalance:
      type: object
      properties:
        availablePoints:
          type: integer
          example: 2500
        monetaryEquivalent:
          type: number
          format: double
          description: Value in GBP based on current redemption rate
          example: 125.0
        pendingPoints:
          type: integer
          description: Points from transactions still processing
          example: 0
        tierMultiplier:
          type: number
          format: double
          example: 1.5
        lastUpdated:
          type: string
          format: date-time
      example:
        availablePoints: 2500
        monetaryEquivalent: 125.0
        pendingPoints: 0
        tierMultiplier: 1.5
        lastUpdated: '2026-05-18T11:45:00Z'
    Transaction:
      type: object
      properties:
        transactionId:
          type: string
          format: uuid
        type:
          type: string
          enum:
          - earn
          - redeem
          - adjustment
          - bonus
          - expired
          - reversal
        points:
          type: integer
          description: Positive for credits, negative for debits
          example: 76
        runningBalance:
          type: integer
          example: 2576
        referenceId:
          type: string
          description: Order ID, adjustment ID, or campaign ID
          example: ORD-2026-001234
        channel:
          type: string
          enum:
          - online
          - in_store
          - app
          - system
        description:
          type: string
          example: Purchase at Next Oxford Street
        createdAt:
          type: string
          format: date-time
      example:
        transactionId: f1e2d3c4-b5a6-7890-fedc-ba0987654321
        type: earn
        points: 76
        runningBalance: 2576
        referenceId: ORD-2026-001234
        channel: online
        description: Online purchase
        createdAt: '2026-05-18T14:30:00Z'
    Tier:
      type: object
      properties:
        tierId:
          type: string
          format: uuid
        name:
          type: string
          example: Gold
        threshold:
          type: integer
          description: Lifetime points required to reach this tier
          example: 5000
        earnRateMultiplier:
          type: number
          format: double
          example: 1.5
        badgeColor:
          type: string
          example: '#FFD700'
        benefits:
          type: array
          items:
            type: string
          example:
          - 1.5x points on all purchases
          - Early access to sales (24 hours)
          - Free delivery on all orders
      example:
        tierId: c3d4e5f6-a7b8-9012-cdef-345678901234
        name: Gold
        threshold: 5000
        earnRateMultiplier: 1.5
        badgeColor: '#FFD700'
        benefits:
        - 1.5x points on all purchases
        - Early access to sales (24 hours)
        - Free delivery on all orders
...
    PaginationMeta:
      type: object
      properties:
        pageNumber:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 156
        totalPages:
          type: integer
          example: 8
        hasNextPage:
          type: boolean
          example: true
        hasPreviousPage:
          type: boolean
          example: false
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
          example: https://loyalty.next.co.uk/problems/validation-error
        title:
          type: string
          description: Short human-readable summary
          example: Validation Error
        status:
          type: integer
          description: HTTP status code
          example: 422
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
          example: The 'email' field must be a valid email address
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence
          example: /api/v1/auth/register
        traceId:
          type: string
          description: Correlation ID for distributed tracing
          example: abc123def456
        errors:
          type: array
          description: Detailed field-level validation errors
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Must be a valid email address
              code:
                type: string
                example: INVALID_FORMAT
      example:
        type: https://loyalty.next.co.uk/problems/validation-error
        title: Validation Error
        status: 422
        detail: One or more fields failed validation
        instance: /api/v1/auth/register
        traceId: abc123def456
        errors:
        - field: email
          message: Must be a valid email address
          code: INVALID_FORMAT
...
  responses:
    BadRequest:
      description: Bad Request — malformed request syntax or invalid parameters
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/bad-request
            title: Bad Request
            status: 400
            detail: Request body is malformed or missing required fields
            traceId: trace-001
    Unauthorized:
      description: Unauthorized — missing or invalid authentication credentials
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/unauthorized
            title: Unauthorized
            status: 401
            detail: Access token is missing, expired, or invalid
            traceId: trace-002
    Forbidden:
      description: Forbidden — authenticated but insufficient permissions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/forbidden
            title: Forbidden
            status: 403
            detail: You do not have permission to perform this action
            traceId: trace-003
    NotFound:
      description: Not Found — requested resource does not exist
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/not-found
            title: Not Found
            status: 404
            detail: The requested resource was not found
            traceId: trace-004
    UnprocessableEntity:
      description: Unprocessable Entity — request understood but validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/validation-error
            title: Validation Error
            status: 422
            detail: One or more fields failed validation
            traceId: trace-005
            errors:
            - field: email
              message: Email already registered
              code: DUPLICATE
    InternalServerError:
      description: Internal Server Error — unexpected server-side failure
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/internal-error
            title: Internal Server Error
            status: 500
            detail: An unexpected error occurred. Please try again later.
            traceId: trace-006